Add files via upload
Dataset and R script used to produce analysis of the relationship between online terrorist activity and real-world outcomes by global jihadist movement (GJM) groups. The dataset contains estimates of GJM groups' online activity from 1998 to 2019, membership estimates for that period, aggregated yearly data on the number of attacks by GJM groups for the period from ACLED and START (for those raw data values, you will need to go to ACLED and START and request those datasets as they cannot be reproduced in full, per terms of service), a list of the GJM groups that data collection was done on, and a sheet of values imported into R for analysis. 

The R script contains the various models and code used for the graphics and tables found in the report. All values that appear in the report can be reproduced with the code in the script.